Changed the definition of first and second Helix when the borders for Helices are calculated.
Applied this change only to rossman3.pl, not to rossman1.pl
Changed the order in a way that firstHelixMin (= min border for a helix) is now the border on the outermost left position when looking at the graph representation. firstHelixMax and secondHelixMin are in the middle, and secondHelixMax is the outermost right one.
I find this to be a more logical numbering as it is in rossman1 right now.