Added function to save the amino acid graph in a new output format
Added new amino acid graph represenation that can be used by the software FANMOD (motif detection).
Therefore added a function to convert the amino acid graph to this represenation and save the graph file to disc.
Since the vertex indices are different from the standard amino acid graph, a second file is saved showing the mapping of the standard AAGraph vertex IDs to the vertex IDs of this new simple represenation, with some added additional information.