Added PDB file converter that reads a PDB file with multiple models and converts those models to separated chains.
The converter is still WIP. As of right now, the format of the converted PDB file is broken, the TER records need to be corrected, and the MASTER entry needs to be added.

Since some biological unit PDB files represent multimeric protein structures with multiple models instead of chains, those models need to be converted to chains because the DSSP algorithm only works on a single model, as does the PPI calculation.