Updated the converting of PDB files with multiple models.
The PDB record layout is now preserved and the MASTER record is added.

Note that the output file may not conform the PDB format but works with the DSSP algorithm and the VPLG software.
Especially the records appearing before the first atom record are simply copied from the original PDB file and not updated.