Update the converting of models to chains.
For PDB files that only include one model the 'MODEL' and 'ENDMDL' lines are removed now anyway. The main reason for this is that the Reduce software will introduce problems when parsing 'MODEL' lines because the whitespace behind those entries will be deleted, thus leading to errors when we try to parse the line later on.